Move `consolePrinter` to its own file to expose it as a public API
